C2C添加

+++ src/main/java/com/zhongqijia/pay/bean/paysand/C2CSandCallBack.java
@@ -29,5 +29,5 @@ public class C2CSandCallBack
 
     private String transType;
 
-    private int userFeeAmt;
+    private double userFeeAmt;
 }
